+++ recipes/cron.rb
@@ -20,8 +20,6 @@
 # limitations under the License.
 #
 
-require 'digest/md5'
-
 # include helper methods
 class ::Chef::Recipe
   include ::Opscode::ChefClient::Helpers
@@ -32,31 +30,16 @@ client_bin = find_chef_client
 node.default['chef_client']['bin'] = client_bin
 create_chef_directories
 
-dist_dir, conf_dir = value_for_platform_family(
-  ['amazon'] => %w( redhat sysconfig ),
-  ['debian'] => %w( debian default ),
-  ['rhel'] => %w( redhat sysconfig ),
-  ['fedora'] => %w( redhat sysconfig ),
-  ['suse'] => %w( suse sysconfig )
-)
-
 # Stop any running chef-client services
 if node['os'] == 'linux'
-  template '/etc/init.d/chef-client' do
-    source "#{dist_dir}/init.d/chef-client.erb"
-    mode '0755'
-    variables(client_bin: client_bin)
-  end
-
-  template "/etc/#{conf_dir}/chef-client" do
-    source "#{dist_dir}/#{conf_dir}/chef-client.erb"
-    mode '0644'
-  end
-
   service 'chef-client' do
     supports status: true, restart: true
-    provider Chef::Provider::Service::Upstart if node['chef_client']['init_style'] == 'upstart'
     action [:disable, :stop]
+    ignore_failure true
+  end
+
+  file '/etc/init.d/chef-client' do
+    action :delete
   end
 end
 
@@ -65,12 +48,12 @@ when 'openindiana', 'opensolaris', 'nexentacore', 'solaris2', 'smartos', 'omnios
   service 'chef-client' do
     supports status: true, restart: true
     action [:disable, :stop]
-    provider Chef::Provider::Service::Solaris
     ignore_failure true
   end
 
 when 'freebsd'
   template '/etc/rc.d/chef-client' do
+    source 'default/freebsd/chef-client.erb'
     owner 'root'
     group 'wheel'
     variables client_bin: client_bin
@@ -87,17 +70,6 @@ when 'freebsd'
   end
 end
 
-# Generate a uniformly distributed unique number to sleep.
-if node['chef_client']['splay'].to_i > 0
-  seed = node['shard_seed'] || Digest::MD5.hexdigest(node.name).to_s.hex
-  sleep_time = seed % node['chef_client']['splay'].to_i
-else
-  sleep_time = nil
-end
-log_file   = node['chef_client']['cron']['log_file']
-append_log = node['chef_client']['cron']['append_log'] ? '>>' : '>'
-daemon_options = " #{node['chef_client']['daemon_options'].join(' ')} " if node['chef_client']['daemon_options'].any?
-
 # If "use_cron_d" is set to true, delete the cron entry that uses the cron
 # resource built in to Chef and instead use the cron_d LWRP.
 if node['chef_client']['cron']['use_cron_d']
@@ -105,20 +77,16 @@ if node['chef_client']['cron']['use_cron_d']
     action :delete
   end
 
-  cron_d 'chef-client' do
-    minute  node['chef_client']['cron']['minute']
-    hour    node['chef_client']['cron']['hour']
-    weekday node['chef_client']['cron']['weekday']
-    path    node['chef_client']['cron']['path'] if node['chef_client']['cron']['path']
-    mailto  node['chef_client']['cron']['mailto'] if node['chef_client']['cron']['mailto']
-    user    'root'
-    cmd = ''
-    cmd << "/bin/sleep #{sleep_time}; " if sleep_time
-    cmd << "#{env_vars} " if env_vars?
-    cmd << "#{node['chef_client']['cron']['nice_path']} -n #{process_priority} " if process_priority
-    cmd << "#{client_bin} #{daemon_options}#{append_log} #{log_file} 2>&1"
-    cmd << ' || echo "Chef client execution failed"' if node['chef_client']['cron']['mailto']
-    command cmd
+  chef_client_cron 'chef-client cron.d job' do
+    minute            node['chef_client']['cron']['minute']
+    hour              node['chef_client']['cron']['hour']
+    weekday           node['chef_client']['cron']['weekday']
+    chef_binary_path  node['chef_client']['cron']['path'] if node['chef_client']['cron']['path']
+    mailto            node['chef_client']['cron']['mailto'] if node['chef_client']['cron']['mailto']
+    splay             node['chef_client']['splay']
+    log_file_name     node['chef_client']['cron']['log_file']
+    append_log_file   node['chef_client']['cron']['append_log']
+    daemon_options    node['chef_client']['daemon_options']
   end
 else
   # Non-linux platforms don't support cron.d so we won't try to remove a cron_d resource.
@@ -128,6 +96,11 @@ else
       action :delete
     end
   end
+
+  sleep_time = splay_sleep_time(node['chef_client']['splay'].to_i)
+  log_file   = node['chef_client']['cron']['log_file']
+  append_log = node['chef_client']['cron']['append_log'] ? '>>' : '>'
+  daemon_options = " #{node['chef_client']['daemon_options'].join(' ')} " if node['chef_client']['daemon_options'].any?
 
   cron 'chef-client' do
     minute  node['chef_client']['cron']['minute']