+++ test/cookbooks/test/recipes/systemd_timer_resource.rb
@@ -1,0 +1,5 @@
+chef_client_systemd_timer 'schedule chef-client to run as cron job' do
+  accept_chef_license true
+  daemon_options ['--run-lock-timeout 0']
+  environment 'FOO' => '1', 'BAR' => '2'
+end