+++ test/integration/cron_resource/cron_spec.rb
@@ -1,0 +1,4 @@
+describe file('/etc/cron.d/chef-client') do
+  its('content') { should match(/chef-client/) }
+  its('content') { should match(/--run-lock-timeout 0/) }
+end